Extracted findOneRecordQuery
