+++ packages/twenty-front/src/modules/object-record/record-show/record-detail-section/components/RecordDetailRelationRecordsListItem.tsx
@@ -185,6 +185,10 @@ export const RecordDetailRelationRecordsListItem = ({
     [isExpanded],
   );
 
+  console.log({
+    relationRecord,
+  });
+
   return (
     <>
       <StyledListItem isDropdownOpen={isDropdownOpen}>
